Consolidate routes into routes/ directory instead of creating routes-elysia/
