+++ routes/cart.ts
@@ -1,60 +1,59 @@
-import express, { Request, Response } from 'express';
+import { Elysia } from 'elysia';
 import db from '../db/index';
-import bodyParser from 'body-parser';
 import decodeJWT from '../utils/decodeJWT';
 import { User, CartItem } from '../types';
 
-const cartRouter = express.Router();
-
-cartRouter.use(bodyParser.json());
-cartRouter.use(
-  bodyParser.urlencoded({
-    extended: true,
+const cartRoutes = new Elysia({ prefix: '/cart' })
+  // GET cart total
+  .get('/total', async ({ headers, set }) => {
+    const token = headers.authorization;
+    const user = decodeJWT(token as string);
+    const selectObject = await db.query<User>('SELECT id FROM users WHERE username = $1', [user]);
+    const userID = selectObject.rows[0].id;
+    const results = await db.query<{ total: string }>('SELECT SUM(carts.amount*products.price) AS total FROM carts, products WHERE user_id = $1 AND products.id=carts.product_id', [userID]);
+    
+    if (!results.rows[0].total) return;
+    const total = results.rows[0].total.slice(1);
+    set.status = 200;
+    return total;
   })
-);
-
-cartRouter.get('/total', async(req: Request, res: Response) => {
-  const token = req.headers.authorization;
-  const user = decodeJWT(token as string); 
-  const selectObject = await db.query<User>('SELECT id FROM users WHERE username = $1', [user]);
-  const userID = selectObject.rows[0].id;
-  const results = await db.query<{ total: string }>('SELECT SUM(carts.amount*products.price) AS total FROM carts, products WHERE user_id = $1 AND products.id=carts.product_id', [userID]);
-  if(!results.rows[0].total) return;
-  const total = results.rows[0].total.slice(1);
-  res.status(200).send(total);
-});
-
-cartRouter.get('/', async(req: Request, res: Response) => {
-  const token = req.headers.authorization;
-  const user = decodeJWT(token as string); 
-  const selectObject = await db.query<User>('SELECT id FROM users WHERE username = $1', [user]);
-  const userID = selectObject.rows[0].id;
-  const results = await db.query<{ id: string; amount: number; name: string }>('SELECT carts.id, carts.amount, products.name FROM carts, products WHERE user_id = $1 AND products.id=carts.product_id;', [userID]);
-  res.status(200).send(results.rows);
-});
-
-cartRouter.put('/:id', (req: Request, res: Response) => {
-  const id = req.params.id;
-  const amount = Number(req.body.amount);
-  db.query<CartItem>('UPDATE carts SET amount = $2 WHERE id = $1 RETURNING *', [id, amount], (error, results) => {
-    if (error) {
+  // GET cart items
+  .get('/', async ({ headers, set }) => {
+    const token = headers.authorization;
+    const user = decodeJWT(token as string);
+    const selectObject = await db.query<User>('SELECT id FROM users WHERE username = $1', [user]);
+    const userID = selectObject.rows[0].id;
+    const results = await db.query<{ id: string; amount: number; name: string }>('SELECT carts.id, carts.amount, products.name FROM carts, products WHERE user_id = $1 AND products.id=carts.product_id;', [userID]);
+    set.status = 200;
+    return results.rows;
+  })
+  // PUT update cart item amount
+  .put('/:id', async ({ params, body, set }) => {
+    const { id } = params;
+    const { amount } = body as any;
+    const numAmount = Number(amount);
+    
+    try {
+      const results = await db.query<CartItem>('UPDATE carts SET amount = $2 WHERE id = $1 RETURNING *', [id, numAmount]);
+      set.status = 200;
+      return results.rows[0];
+    } catch (error) {
       console.log('error');
       throw error;
     }
-    res.status(200).json(results.rows[0]);
-  });
-});
-
-cartRouter.delete('/:id', (req: Request, res: Response) => {
-  const id = req.params.id;
-
-  db.query('DELETE FROM carts WHERE id = $1', [id], (error, results) => {
-    if (error) {
+  })
+  // DELETE cart item
+  .delete('/:id', async ({ params, set }) => {
+    const { id } = params;
+    
+    try {
+      await db.query('DELETE FROM carts WHERE id = $1', [id]);
+      set.status = 200;
+      return;
+    } catch (error) {
       console.log('error');
       throw error;
     }
-    res.status(200).send();
   });
-});
 
-export default cartRouter;
+export default cartRoutes;
