+++ routes/registration.ts
@@ -1,20 +1,9 @@
-import express, { Request, Response, NextFunction } from 'express';
+import { Elysia, t } from 'elysia';
 import db from '../db/index';
-import bodyParser from 'body-parser';
 import bcrypt from 'bcrypt';
 import validator from 'validator';
-import { check, validationResult } from 'express-validator';
 import generateToken from '../utils/generateToken';
-import { User, RegisterRequest } from '../types';
-
-const registerRouter = express.Router();
-
-registerRouter.use(bodyParser.json());
-registerRouter.use(
-  bodyParser.urlencoded({
-    extended: true,
-  })
-);
+import { User } from '../types';
 
 const passwordHasher = async (password: string, saltRounds: number): Promise<string | null> => {
   try {
@@ -27,65 +16,87 @@ const passwordHasher = async (password: string, saltRounds: number): Promise<str
   return null;
 };
 
-registerRouter.post('/', [
-  check('password').isLength({ max: 20, min: 5 }),
-  check('email').isEmail()], async (req: Request, res: Response, next: NextFunction) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.status(422).json({ errors: errors.array() });
-  }
+// Email validation function
+const isEmail = (email: string): boolean => {
+  return validator.isEmail(email);
+};
 
-  const selectObject = await db.query<{ count: string }>('SELECT COUNT(*) FROM users');
-  const totalUsers = Number(selectObject.rows[0].count);
-  const id = totalUsers + 1;
+// Password validation function
+const isValidPassword = (password: string): boolean => {
+  return password.length >= 5 && password.length <= 20;
+};
 
-  let first = req.body.first; 
-  let last = req.body.last; 
-  let email = req.body.email; 
-  let username = req.body.username; 
-  let password = req.body.password;
+const registerRoutes = new Elysia({ prefix: '/register' })
+  .post('/', async ({ body, set }) => {
+    // Type the body
+    const { first, last, email, username, password } = body as any;
 
-  if (!first || !last || !email || !username || !password) {
-    return res.status(400).send({ error: true, message: "Missing one or more required fields"});
-  }
+    // Validate required fields
+    if (!first || !last || !email || !username || !password) {
+      set.status = 400;
+      return { error: true, message: "Missing one or more required fields" };
+    }
 
-  first = validator.escape(first); 
-  last = validator.escape(last); 
-  email = validator.escape(email); 
-  username = validator.escape(username); 
-  password = validator.escape(password);
+    // Validate email format
+    if (!isEmail(email)) {
+      set.status = 422;
+      return { errors: [{ field: 'email', message: 'Invalid email format' }] };
+    }
 
-  const emailExistsResult = await db.query<User>('SELECT * FROM users WHERE email = $1', [email]);
-  const emailExists = emailExistsResult.rows;
-  if (emailExists?.length) {
-    console.log("email already exists!");
-    return res.status(400).send({ error: true, message: "The provided email is linked to an existing account."});
-  }
+    // Validate password length
+    if (!isValidPassword(password)) {
+      set.status = 422;
+      return { errors: [{ field: 'password', message: 'Password must be between 5 and 20 characters' }] };
+    }
 
-  try {
-    const selectText = 'SELECT * FROM users WHERE username = $1';
-    const values = [username];
-    await db.query<User>(selectText, values, async (error, results) => {
-      if (error) {
-        console.log('error');
-        throw error;
-      }
+    // Sanitize inputs
+    const sanitizedFirst = validator.escape(first);
+    const sanitizedLast = validator.escape(last);
+    const sanitizedEmail = validator.escape(email);
+    const sanitizedUsername = validator.escape(username);
+    const sanitizedPassword = validator.escape(password);
+
+    // Check if email exists
+    const emailExistsResult = await db.query<User>('SELECT * FROM users WHERE email = $1', [sanitizedEmail]);
+    const emailExists = emailExistsResult.rows;
+    if (emailExists?.length) {
+      console.log("email already exists!");
+      set.status = 400;
+      return { error: true, message: "The provided email is linked to an existing account." };
+    }
+
+    try {
+      // Check if username exists
+      const selectText = 'SELECT * FROM users WHERE username = $1';
+      const values = [sanitizedUsername];
+      const results = await db.query<User>(selectText, values);
       const user = results.rows[0];
+      
       if (user) {
         console.log("User already exists!");
-        return res.status(400).send({ error: true, message: "Username already exists. Please choose another username"});
+        set.status = 400;
+        return { error: true, message: "Username already exists. Please choose another username" };
       }
-      const insertText = 'INSERT INTO users (id, first_name, last_name, email, username, password)VALUES($1, $2, $3, $4, $5, $6) RETURNING *';
-      const hashedPassword = await passwordHasher(password, 10);
-      const insertValues = await [id, first, last, email, username, hashedPassword];
-      const newUser = await db.query<User>(insertText, insertValues);
-      const newUsername = await newUser.rows[0].username;
-      const token = generateToken({ username: newUsername });
-      return res.status(200).send({ error: false, token, message: "User created" });
-    }); 
-  } catch (err: any) {
-    res.status(500).json({ message: err.message });
-  }
-});
 
-export default registerRouter;
+      // Get total users to generate new ID
+      const selectObject = await db.query<{ count: string }>('SELECT COUNT(*) FROM users');
+      const totalUsers = Number(selectObject.rows[0].count);
+      const id = totalUsers + 1;
+
+      // Hash password and insert user
+      const insertText = 'INSERT INTO users (id, first_name, last_name, email, username, password)VALUES($1, $2, $3, $4, $5, $6) RETURNING *';
+      const hashedPassword = await passwordHasher(sanitizedPassword, 10);
+      const insertValues = [id, sanitizedFirst, sanitizedLast, sanitizedEmail, sanitizedUsername, hashedPassword];
+      const newUser = await db.query<User>(insertText, insertValues);
+      const newUsername = newUser.rows[0].username;
+      const token = generateToken({ username: newUsername });
+      
+      set.status = 200;
+      return { error: false, token, message: "User created" };
+    } catch (err: any) {
+      set.status = 500;
+      return { message: err.message };
+    }
+  });
+
+export default registerRoutes;
